feat(admin/file-types): merge add & new page's into one
