+++ apps/web/src/app/[lang]/(main)/(upwithcrowd)/management/file/file-types/[fileTypeId]/_components/form-new.tsx
@@ -1,48 +1,41 @@
 "use client";
 
 import {
-  $UpwithCrowd_FileTypes_FileTypeUpdateDto,
+  $UpwithCrowd_FileTypes_FileTypeDto,
   type UpwithCrowd_FileTypeGroup_FileTypeGroupListDto,
-  type UpwithCrowd_FileTypes_FileTypeUpdateDto,
-  type UpwithCrowd_FileTypes_ListFileTypeDto,
+  type UpwithCrowd_FileTypes_FileTypeDto,
   type UpwithCrowd_Providers_ProvidersListDto,
 } from "@ayasofyazilim/upwithcrowd-saas/UPWCService";
-import {putFileTypeApi} from "@repo/actions/upwithcrowd/file-type/put-actions";
+import {postFileTypeApi} from "@repo/actions/upwithcrowd/file-type/post-actions";
 import {SchemaForm} from "@repo/ayasofyazilim-ui/organisms/schema-form";
 import {CustomComboboxWidget} from "@repo/ayasofyazilim-ui/organisms/schema-form/widgets";
 import {handlePutResponse} from "@repo/utils/api";
 import {useRouter} from "next/navigation";
 import type {IdentityServiceResource} from "src/language-data/core/IdentityService";
 
-export default function Form({
+export default function NewForm({
   languageData,
-  fileTypeData,
   fileTypeGroupData,
   providerData,
 }: {
   languageData: IdentityServiceResource;
-  fileTypeData: UpwithCrowd_FileTypes_ListFileTypeDto;
   fileTypeGroupData: UpwithCrowd_FileTypeGroup_FileTypeGroupListDto[];
+
   providerData: UpwithCrowd_Providers_ProvidersListDto[];
 }) {
   const router = useRouter();
   return (
-    <SchemaForm<UpwithCrowd_FileTypes_FileTypeUpdateDto>
+    <SchemaForm<UpwithCrowd_FileTypes_FileTypeDto>
       className="flex flex-col gap-4"
-      formData={{
-        ...fileTypeData,
-        fileTypeGroupNamespace: fileTypeGroupData.find((item) => item.id === fileTypeData.fileTypeGroupID)?.namespace,
-      }}
       onSubmit={({formData}) => {
         if (!formData) return;
-        void putFileTypeApi({
-          id: fileTypeData.id,
+        void postFileTypeApi({
           requestBody: formData,
         }).then((res) => {
           handlePutResponse(res, router);
         });
       }}
-      schema={$UpwithCrowd_FileTypes_FileTypeUpdateDto}
+      schema={$UpwithCrowd_FileTypes_FileTypeDto}
       submitText={languageData.Save}
       uiSchema={{
         fileTypeGroupNamespace: {
